Remove Gst.deinit (proactive bug fix)
Gst.deinit caused the Dino proces to linger in the background when I
tried using Gst in the phone ringer plugin. My reasoning for leaving it
out even though Gst is not in use anymore in the ringer plugin is that
this is a nasty bug that might crop up again in the future when someone
tries to do anything else at all with Gst.

I did the following things:

- checked this https://gstreamer.freedesktop.org/documentation/gstreamer/gst.html#gst_deinit
- tested that Dino works without the Gst.deinit
- tried looking for a specific reason for the deinit with git log -L 276,+10 -- plugins/rtp/src/plugin.vala

I didn't find anything so this made me conclude that it is better to
leave it out.